Refactoring of base consepts
- renamed base Attribute to StepDefinition
- renamed all StepDefinitions (should be backwards compatible)
- Moved step definitions into a seperate folder
