+++ source/Xunit.Gherkin.Quick/StepDefinition/And.cs
@@ -1,0 +1,11 @@
+﻿namespace Xunit.Gherkin.Quick
+{
+
+    public sealed class And : StepDefinitionAttributeBase
+    {
+        public And(string pattern)
+            : base("And", pattern)
+        {
+        }
+    }
+}
