+++ source/Xunit.Gherkin.Quick/StepDefinition/Given.cs
@@ -1,0 +1,11 @@
+﻿namespace Xunit.Gherkin.Quick
+{
+    public sealed class Given : StepDefinitionAttributeBase
+    {
+        public Given(string pattern)
+            : base("Given", pattern)
+        {
+
+        }
+    }
+}
