+++ source/Xunit.Gherkin.Quick/StepDefinition/GivenWhenThen.cs
@@ -1,0 +1,12 @@
+﻿namespace Xunit.Gherkin.Quick
+{
+
+    public sealed class GivenWhenThen : StepDefinitionAttributeBase
+    {
+        public GivenWhenThen(string pattern)
+            : base("But", pattern)
+        {
+        }
+    }
+
+}
