+++ source/Xunit.Gherkin.Quick/StepDefinition/Then.cs
@@ -1,0 +1,12 @@
+﻿namespace Xunit.Gherkin.Quick
+{
+
+    public sealed class Then : StepDefinitionAttributeBase
+    {
+        public Then(string pattern)
+            : base("Then", pattern)
+        {
+
+        }
+    }
+}
