+++ source/Xunit.Gherkin.Quick/StepDefinition/When.cs
@@ -1,0 +1,12 @@
+﻿namespace Xunit.Gherkin.Quick
+{
+
+    public sealed class When : StepDefinitionAttributeBase
+    {
+        public When(string pattern)
+            : base("When", pattern)
+        {
+
+        }
+    }
+}
